Update cmake build scripts.

+++ bin/cmake-clean.sh
@@ -2,7 +2,7 @@
 set -o errexit -o nounset -o pipefail
 binary_dir=$(cat cmake-binary-dir.txt)
 if [[ -e Makefile ]]; then make clean || true; fi
-if [[ $binary_dir != '.' ]]; then
+if [[ $binary_dir != '.' &&  $binary_dir != '..' ]]; then
   rm -fr $binary_dir
 else
   rm -f  $binary_dir/cmake_install.cmake
